Remove ids HashSet from Schedule
Previously, the we used Schedule::ids to track the set of coroutines that had not yet been cancelled. AsyncRuntime::inactive_coroutines can already serve this purpose, so we remove Schedule::Ids and move the checking for a canceled coroutine to AsyncRuntime::poll. These functions are called the most frequently, so it makes sense to reduce the number of HashSets and HashMaps in use.