Remove AsyncRuntime::active_coroutines
Here we sacrifice some error reporting for a couple more microseconds of latency.

AsyncRuntime::active_coroutines's only purpose was in AsyncRuntime::coroutine_status(). When given an invalid coroutine id, the previous implementation of coroutine_status() would panic. Here, we just default to CoroutineStatus::Active. The behavior in correct operation should be the same. What we get in return is the removal of another HashSet from the critical path.