[inetstack] Workaround: Push Empty Buffer on FIN
This is a workaround for issue
https://github.com/demikernel/demikernel/issues/369.

Currently, our TCP stack misses to inform the socket layer that a
connetion is closing (ie. a FIN has arrived).

This commit works around this issue and pushes a zero-length buffer
when a FIN has been received and the connection is in a closing state.