[dpdk-rs] Workaround: redefine RTE_TOOLCHAIN_XXX for host toolchain
there is a bug in dpdk headers where the toolchain preprocessor
definition used to build dpdk is retained and then incorrectly directs
conditional compilation during application builds.

redefine the RTE_TOOLCHAIN_XXX macros as appropriate using the host toolchain
predefined macros to get correct conditional evaluation when consuming dpdk
headers.

this is just a workaround for now, i will work with the dpdk community
to upstream a fix which will eventually allow this change to be removed.

Signed-off-by: Tyler Retzlaff <roretzla@microsoft.com>